packages/ragbits-evaluate/pyproject.toml
[project]
name = "ragbits-evaluate"
version = "0.3.0"
description = "Evaluation module for Ragbits components"
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
authors = [
    { name = "deepsense.ai", email = "ragbits@deepsense.ai"}
]
keywords = [
    "Retrieval Augmented Generation",
    "RAG",
    "Large Language Models",
    "LLMs",
    "Generative AI",
    "GenAI",
    "Evaluation"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Natural Language :: English",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = ["hydra-core~=1.3.2", "neptune~=1.12.0", "ragbits-core==0.3.0", "optuna==4.0.0", "distilabel==1.4.1"]

[project.optional-dependencies]
relari = [
    "continuous-eval~=0.3.12",
]

[tool.uv]
dev-dependencies = [
    "pre-commit~=3.8.0",
    "pytest~=8.3.3",
    "pytest-cov~=5.0.0",
    "pytest-asyncio~=0.24.0",
    "pip-licenses>=4.0.0,<5.0.0"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/ragbits"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
